refactor: images clean up
