+++ src/components/ApplicationGroup/Details/TriggerView/BulkCDTrigger.tsx
@@ -56,7 +56,7 @@ import { ReactComponent as PlayIcon } from '@Icons/ic-play-outline.svg'
 import { ReactComponent as Error } from '@Icons/ic-warning.svg'
 import { ReactComponent as UnAuthorized } from '@Icons/ic-locked.svg'
 import { ReactComponent as Tag } from '@Icons/ic-tag.svg'
-import emptyPreDeploy from '../../../../assets/img/empty-pre-deploy.png'
+import emptyPreDeploy from '../../../../assets/img/empty-pre-deploy.webp'
 import notAuthorized from '../../../../assets/img/ic-not-authorized.svg'
 import CDMaterial from '../../../app/details/triggerView/cdMaterial'
 import { BulkSelectionEvents, MATERIAL_TYPE, RuntimeParamsErrorState } from '../../../app/details/triggerView/types'
