+++ src/components/ApplicationGroup/Details/TriggerView/BulkCITrigger.tsx
@@ -53,9 +53,9 @@ import { ReactComponent as ICError } from '../../../../assets/icons/ic-alert-tri
 import { ReactComponent as Storage } from '../../../../assets/icons/ic-storage.svg'
 import { ReactComponent as OpenInNew } from '../../../../assets/icons/ic-open-in-new.svg'
 import { ReactComponent as InfoIcon } from '../../../../assets/icons/info-filled.svg'
-import externalCiImg from '../../../../assets/img/external-ci.png'
-import linkedCDBuildCIImg from '../../../../assets/img/linked-cd-bulk-ci.png'
-import linkedCiImg from '../../../../assets/img/linked-ci.png'
+import externalCiImg from '../../../../assets/img/external-ci.webp'
+import linkedCDBuildCIImg from '../../../../assets/img/linked-cd-bulk-ci.webp'
+import linkedCiImg from '../../../../assets/img/linked-ci.webp'
 import { getModuleConfigured } from '../../../app/details/appDetails/appDetails.service'
 import { DOCUMENTATION, SOURCE_NOT_CONFIGURED, URLS, ViewType } from '../../../../config'
 import MaterialSource from '../../../app/details/triggerView/MaterialSource'
